fix: wrong incoming bb referenced on phi nodes
on phi nodes, the incoming basic blocks must be
referenced correctly; current basic blocks were
not tracked at all. when a basic block was
switched during regular codegen, the upcoming
phi node was generated with the refs of the
original basic blocks.

(imagine an "if", which encapsulated another "if"
on one of its branches, the second if will "switch"
basic blocks, whilst the first will just use the
"original" basic block of the immediate branch)