provide pint impl of =
we need to do some trickery to make
this work in interpreter and compiler
with just one version of the source.

for now i will leave this implementations
in the "helloworld.pint" file.

once the compiler is complete enough to
compile itself, i will merge these implementations
into the compiler source itself.