split out std.pint and parser.pint from pint.pint
because of the includer hack, we have another
chicken/egg problem: to get includer compiled
we need pint.pint with includes. bootstrap hack
is provided via Makefile.

because includer hack must now be executed on
pint.pint as well as the to be compiled script,
and because we cannot provide _both_ transformed
inputs via stdin we use process substitution in
Makefile and tmpfiles in build-pint.