stdlib: provide polyfill for str
when writing this code a problem in the generated code was found
if you use e.g. printf, the compiler will not generate its vaargs
signature, thus if printf is used with different signatures, the
compiler will generate invalid code.

furthermore even if the signatures match, there seems to be some
problem calling vaargs function with a non vaarg signature:

when we tried to define printf(i8*,i8*) and call it with a
'bitcasted' double as its second argument, it did not work
reliably.

thus for now, strings are appended using strncpy and numbers
are formatted via snprintf