get rid of extern.cc inline-fn expose hack
LLVMInitializeNative* fns are declared as
static inline. Hence to use them, they need
to be present in the same translation unit as
the compiled pint code. since this is not going
to happen, we now use the proper initialization
fns for x86(64), which are available via
ordinary 'external' link level.

we can now drop the last bit of cc code :)