put current module to generate code for into state
for macro execution we need to generate code also into the
to be jitted module. thus we need to be able to do codegen
for a given module (once for emitted binary, once for to be
executed macro).

the "current module" is packed into the state, code gen will
generate code into this module. (indeed the current module
was already packed via the current basic block, we use this
to determine the parent module of the block, i.e. the current
module)