macro form always evals to its statically evaled value
but there is an important limitation:

due to the current design we need to track separate state
when generating code and jitting code (because we have
separate modules for translating and jitting).

this could be unified in two ways, using llvm-orc (and
use the very same code that is used for executable output)

or the other way round, use the modules used for jitting
also for output, but then output will consist of multiple
modules.

this state separation makes implementing macros more
complicated than it need to be. for now they are
implemented with the limitation, that macro _definitions_
may only occur on the top level.